+++ runtime/streamlib-ipc-types/Cargo.toml
@@ -12,6 +12,7 @@ description = "Shared iceoryx2 payload types for StreamLib IPC"
 
 [dependencies]
 iceoryx2 = "0.8.1"
+tracing.workspace = true
 
 [lints]
 workspace = true
